refactor(storage): delete PropertyStorage::Compact; split column_store into a module dir
The enum is now {Map, Columnar} = transient vs real, documented at the
module header. Compact's three producers: the .kgl load path's transient
Map->Compact rewrite (deleted; schema derivation survives as
rebuild_type_schemas), dead new_compact constructors (deleted), and
disable_columnar — which now de-columnarizes into Map, keeping its four
pinned behaviors without resurrecting the variant. New public primitive
rebuild_columns_to_heap serves vacuum and unspill (unspill lost its
disable/enable round-trip). PropertyStorage:: sites 144->114, ::Compact
23->3 (doc comments); property_storage.rs 526->368 ln. column_store.rs
(2801 ln, over ceiling) -> column_store/{mod,typed_column,tests}.rs,
mechanical move, module path unchanged; path-keyed lint-allowance
identities re-keyed. RDF loader stays Map-transient by evidence;
embedding_carry was test-only all along. API baselines: -4/+3 symbols.
